better

+++ src/shammodels/common/include/shammodels/common/timestep_report.hpp
@@ -18,6 +18,7 @@
 
 #include "shambase/aliases_float.hpp"
 #include "shambase/aliases_int.hpp"
+#include <optional>
 #include <string>
 
 namespace shammodels {
@@ -32,6 +33,6 @@ namespace shammodels {
         f64 alloc_time_host,
         size_t max_mem_device,
         size_t max_mem_host,
-        std::optional<u64> rank_energy_consummed = std::nullopt);
+        std::optional<f64> rank_energy_consummed = std::nullopt);
 
 } // namespace shammodels
